메가 커피, 빽다방 크롤링 서비스 수정

+++ src/main/java/com/example/coffeeproject/model/ATwosomePlace.java
@@ -20,7 +20,7 @@ public class ATwosomePlace {
     private int twosome_id;
 
     private int ounce; // 중량 1oz=29.5ml
-    private int price; // 가격
+    private double price; // 가격
     private String name; // 이름
     private String eng_name; // 영어 이름
     private String note; // 비고
